backend/requirements.txt
arxiv>=2.1.0
httpx>=0.27.0
pydantic>=2.0
openreview-py>=1.57.0
